+++ Cargo.toml
@@ -12,7 +12,7 @@ members  = [
   "crates/swc_remove_export",
   "crates/swc_optimize_barrel",
   "crates/plugin_specilize_module_name",
-  # "crates/swc_plugin_change_package_import",
+  "crates/swc_change_package_import",
 ]
 resolver = "2"
 
@@ -83,6 +83,7 @@ swc_html            = { version = "=0.135.20" }
 swc_html_minifier   = { version = "=0.132.20" }
 swc_node_comments   = { version = "=0.20.12" }
 tikv-jemallocator   = { version = "=0.5.4", features = ["disable_initial_exec_tls"] }
+testing             = { version = "=0.35.14" }
 
 [profile.dev]
 codegen-units = 16      # debug build will cause runtime panic if codegen-unints is default
