+++ crates/binding_options/src/options/js_loader.rs
