+++ crates/binding_options/src/options/raw_module.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
-use loader_compilation::COMPILATION_LOADER_IDENTIFIER;
 use loader_barrel::BARREL_LOADER_IDENTIFIER;
+use loader_compilation::COMPILATION_LOADER_IDENTIFIER;
 use rspack_core::BoxLoader;
 use rspack_loader_react_refresh::REACT_REFRESH_LOADER_IDENTIFIER;
 use rspack_loader_swc::SWC_LOADER_IDENTIFIER;
