+++ crates/loader_barrel/src/lib.rs
@@ -12,13 +12,13 @@ use rspack_core::{
   DependencyCategory, LoaderRunnerContext, ResolveOptionsWithDependencyType, ResolveResult,
   Resolver,
 };
-use serde::Deserialize;
 use rspack_error::{error, AnyhowError, Result};
 use rspack_loader_runner::{Content, Identifiable, Identifier, Loader, LoaderContext};
 use rspack_plugin_javascript::{
   ast::{self, SourceMapConfig},
   TransformOutput,
 };
+use serde::Deserialize;
 use swc_compiler::{IntoJsAst, SwcCompiler};
 use swc_core::{
   base::config::{Options, OutputCharset},
@@ -188,9 +188,15 @@ async fn get_barrel_map(
           .resolve(&file.parent().unwrap().to_path_buf(), &real_req)
           .map_err(|err| error!("Failed to resolve {err:?}"))?;
         if let ResolveResult::Resource(resource) = wildcard_resolve {
-          let res =
-            get_barrel_map_boxed(visited.clone(), resolver.clone(), resource.path, cache_dir.clone(), true, None)
-              .await?;
+          let res = get_barrel_map_boxed(
+            visited.clone(),
+            resolver.clone(),
+            resource.path,
+            cache_dir.clone(),
+            true,
+            None,
+          )
+          .await?;
           if let Some(TransformMapping {
             export_list: sub_export_list,
             wildcard_exports: _,
@@ -221,7 +227,14 @@ fn get_barrel_map_boxed(
   is_wildcard: bool,
   source: Option<String>,
 ) -> Pin<Box<dyn Future<Output = Result<Option<TransformMapping>>> + Send>> {
-  Box::pin(get_barrel_map(visited, resolver, file, cache_dir, is_wildcard, source))
+  Box::pin(get_barrel_map(
+    visited,
+    resolver,
+    file,
+    cache_dir,
+    is_wildcard,
+    source,
+  ))
 }
 
 #[async_trait::async_trait]
