+++ crates/loader_compilation/src/transform.rs
@@ -7,8 +7,8 @@ use swc_core::common::chain;
 use swc_core::ecma::{transforms::base::pass::noop, visit::Fold};
 use swc_env_replacement::env_replacement;
 use swc_keep_export::keep_export;
-use swc_remove_export::remove_export;
 use swc_named_import_transform::{named_import_transform, TransformConfig};
+use swc_remove_export::remove_export;
 
 macro_rules! either {
   ($config:expr, $f:expr) => {
@@ -93,14 +93,11 @@ pub(crate) fn transform<'a>(
   feature_options: &TransformFeatureOptions,
 ) -> impl Fold + 'a {
   chain!(
-    either!(
-      feature_options.optimize_import,
-      |options: &Vec<String>| {
-        named_import_transform(TransformConfig {
-          packages: options.clone(),
-        })
-      }
-    ),
+    either!(feature_options.optimize_import, |options: &Vec<String>| {
+      named_import_transform(TransformConfig {
+        packages: options.clone(),
+      })
+    }),
     either!(
       Some(&vec!["@uni/env".to_string(), "universal-env".to_string()]),
       |options: &Vec<String>| { env_replacement(options.clone()) }
