+++ crates/swc_change_package_import/src/lib.rs
@@ -1,5 +1,5 @@
-mod config;
 mod change_package_import;
+mod config;
 
-pub use config::*;
 pub use change_package_import::*;
+pub use config::*;
