+++ crates/swc_change_package_import/tests/fixture/ice_alias_transform/output.js
