+++ crates/swc_change_package_import/tests/fixture/ice_alias_with_as_transform/output.js
