+++ crates/swc_change_package_import/tests/fixture/ice_as_transform/output.js
