+++ crates/swc_change_package_import/tests/fixture/ice_basic_transform/input.js
