+++ crates/swc_change_package_import/tests/fixture/ice_multiple_transform/output.js
