+++ crates/swc_change_package_import/tests/fixture/multi_literal_transform/output.js
