+++ crates/swc_change_package_import/tests/fixture/multi_specific_transform/output.js
