+++ crates/swc_change_package_import/tests/fixture/single_literal_transform/input.js
