--- crates/swc_plugin_change_package_import/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "swc_plugin_change_package_import"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[profile.release]
-lto = true
-
-[dependencies]
-serde = { workspace = true }
-swc_core = { workspace = true, features = [
-    "ecma_plugin_transform",
-    "ecma_ast",
-    "common",
-    "ecma_utils",
-] }
-
-[dev-dependencies]
-testing = "0.35.11"
